高光谱code/requirements.txt
torch
numpy
scikit-image
Pillow
Django
djangorestframework
scipy
opencv-python